+++ src/js/Tools.js
